rename i2c_lcd since it's no longer esp8266-specific
